refactor: Remove SQL script for adding propietario_id to jugadores table
This commit deletes the SQL script `add_propietario_id_to_jugadores.sql`, which was previously used to add the propietario_id column to the jugadores table along with associated RLS policies and indexing. The removal is part of a refactor to streamline database management and may indicate a shift in how player ownership is handled in the application.